pyproject.toml
[tool.poetry]
name = "sandbox-python"
version = "0.0.1"
description = "A GitHub template repository for Python"
authors = ["ks6088ts <ks6088ts@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
typer = "^0.12.5"
fastapi = {extras = ["standard"], version = "^0.115.0"}
azure-functions = "^1.21.0"
streamlit = "^1.38.0"
neo4j = "^5.25.0"
langchain-experimental = "^0.3.2"
onnxruntime-genai = "^0.4.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
black = "^24.8.0"
isort = "^5.13.2"
ruff = "^0.6.8"
pytest = "^8.3.3"
coverage = "^7.6.1"
pytest-cov = "^5.0.0"


[tool.poetry.group.llms.dependencies]
langchain = "^0.3.1"
langchain-openai = "^0.2.1"
langchain-chroma = "^0.1.4"
langchain-community = "^0.3.1"
langgraph = "^0.2.28"
beautifulsoup4 = "^4.12.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["D203"]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.pytest.ini_options]
addopts = "-ra --cov"
testpaths = ["tests"]
pythonpath = ['.']

[tool.coverage.report]
show_missing = true
